backend/requirements.txt
# Core dependencies
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
pydantic>=2.4.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-multipart>=0.0.6,<0.1.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-dotenv>=1.0.0,<2.0.0
aiofiles>=23.2.0,<24.0.0
httpx>=0.25.0,<0.26.0
requests>=2.31.0,<3.0.0

# Database dependencies
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0,<0.30.0
alembic>=1.12.0,<2.0.0

# Redis and caching
redis>=5.0.0,<6.0.0
hiredis>=2.2.0,<3.0.0

# Security dependencies
cryptography>=41.0.0,<42.0.0
sentry-sdk[fastapi]>=1.38.0,<2.0.0
bleach>=6.1.0,<7.0.0
python-magic>=0.4.27,<0.5.0
PyJWT>=2.8.0,<3.0.0

# Additional security libraries
validators>=0.22.0,<0.23.0
email-validator>=2.1.0,<3.0.0

# Vector and context management
pgvector>=0.2.4,<0.3.0
tiktoken>=0.5.2,<0.6.0

# LlamaIndex and AI dependencies
llama-index>=0.9.30,<0.10.0
llama-index-core>=0.9.30,<0.10.0
llama-index-llms-anthropic>=0.1.7,<0.2.0
llama-index-embeddings-openai>=0.1.6,<0.2.0
llama-index-vector-stores-postgres>=0.1.3,<0.2.0
llama-index-readers-file>=0.1.5,<0.2.0
llama-index-agent-openai>=0.1.5,<0.2.0

# AI Model SDKs
anthropic>=0.18.1,<0.19.0
openai>=1.12.0,<2.0.0

# Data processing
numpy>=1.26.0,<2.0.0
pandas>=2.1.0,<3.0.0
scipy>=1.11.0,<2.0.0

# Additional utilities
tenacity>=8.2.0,<9.0.0
typing-extensions>=4.8.0,<5.0.0
nest-asyncio>=1.5.0,<2.0.0